Ready for initial integration attempt. This is going to be bad.
