+++ locations.c
@@ -81,7 +81,7 @@ public class Location {
 */   
           
        
-int Location_Update(int token)
+int Location_Update(int token, int room)
 {
